requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
rich>=13.7.0

# Web scraping
scrapy>=2.11.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
selenium>=4.15.0
requests-html>=0.10.0

# Financial data scraping
yfinance>=0.2.28
finvizfinance>=0.14.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# LLM and AI
openai>=1.6.0
anthropic>=0.8.0
tiktoken>=0.5.0

# Async processing
aiohttp>=3.9.0
asyncio>=3.4.3

# Database (optional)
sqlalchemy>=2.0.0
alembic>=1.13.0

# Reddit scraping
praw>=7.7.1

# Logging
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.66.0

# Optional: Social media APIs
tweepy>=4.14.0  # Twitter API
python-telegram-bot>=20.0  # Telegram
discord.py>=2.3.0  # Discord